Add mockup data for posts in homepage
Mockup data come from a json file generated with a script to simulate
posts in a database. It has many fields, and one image url at the
moment.

Use json-server as a dev-dependency to serve the posts.json file as
a fake rest api endpoint. The server can be run by doing "npm run server"
in a new terminal. The server runs on port 3001 to prevent conflict.

Size of image containers was fixed to 300px height and set the object-fit
to cover the available space.