Add map and post listing
Uses clustering with supercluster library
Show posts in the map depending on the map boundaries. If user moves
around the map the listing of posts changes
Markers in the map are grouped in clusters if there are enough of
them within an area.

The fetching of the maps occurs at the app level, fetching also changes
depending if boundaries exist or not. So if people is not looking at the
map, a normal limited fetch is asked to the server.

Server needs to look up the database when queried with bounds query
params. If no query params are sent, a normal limited fetch is made.